requirements.txt
# Framework Web
Flask==2.3.0
Werkzeug==2.3.0

# Base de Datos
Flask-SQLAlchemy==3.0.0
SQLAlchemy==2.0.0
pg8000==1.30.3

# API REST y Documentación
flask-restx==1.1.0

# Autenticación JWT
PyJWT==2.8.0

# CORS
Flask-CORS==4.0.0

# OAuth
Authlib==1.2.0

# Email con Brevo (reemplaza Flask-Mail)
sib-api-v3-sdk==7.6.0

# Variables de Entorno
python-dotenv==1.0.0

# Requests HTTP
requests==2.31.0

# Utilidades
python-dateutil==2.8.2

# Servidor WSGI para producción (Render)
gunicorn==21.2.0